+++ osgar/drivers/test_logsocket.py
@@ -2,7 +2,7 @@ import unittest
 from unittest.mock import patch, MagicMock
 
 from osgar.drivers.logsocket import LogTCP
-from osgar.drivers.bus import BusHandler
+from osgar.bus import BusHandler
 
 
 class LogSocketTest(unittest.TestCase):
