+++ osgar/drivers/test_simulator.py
@@ -2,7 +2,7 @@ import unittest
 import math
 
 from osgar.drivers.simulator import SpiderSimulator
-from osgar.drivers.bus import BusHandler
+from osgar.bus import BusHandler
 
 
 class SpiderSimulatorTest(unittest.TestCase):
